requirements.txt
flask==2.2.2
flask_cors==3.0.10
tensorflow==2.2.0
Pillow==9.2.0